# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Config Service",
        "title": "Config Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1323",
    "basePath": "/api",
    "paths": {
...
        "/value-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all value types",
                "produces": [
                    "application/json"
                ],
                "summary": "Get value types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/valuetype.ValueTypeDto"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/value-types/{value_type_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a value type by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get value type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Value type ID",
                        "name": "value_type_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/valuetype.ValueTypeDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.ValueTypeKind": {
            "type": "string",
            "enum": [
                "string",
                "integer",
                "decimal",
                "boolean",
                "json"
            ],
            "x-enum-varnames": [
                "ValueTypeKindString",
                "ValueTypeKindInteger",
                "ValueTypeKindDecimal",
                "ValueTypeKindBoolean",
                "ValueTypeKindJson"
            ]
        },
        "db.ValueValidatorType": {
            "type": "string",
            "enum": [
                "required",
                "min_length",
                "max_length",
                "min",
                "max",
                "min_decimal",
                "max_decimal",
                "regex",
                "json_schema",
                "valid_json",
                "valid_integer",
                "valid_decimal",
                "valid_regex"
            ],
            "x-enum-varnames": [
                "ValueValidatorTypeRequired",
                "ValueValidatorTypeMinLength",
                "ValueValidatorTypeMaxLength",
                "ValueValidatorTypeMin",
                "ValueValidatorTypeMax",
                "ValueValidatorTypeMinDecimal",
                "ValueValidatorTypeMaxDecimal",
                "ValueValidatorTypeRegex",
                "ValueValidatorTypeJsonSchema",
                "ValueValidatorTypeValidJson",
                "ValueValidatorTypeValidInteger",
                "ValueValidatorTypeValidDecimal",
                "ValueValidatorTypeValidRegex"
            ]
        },
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "validation.ValueValidatorParameterType": {
            "type": "string",
            "enum": [
                "none",
                "integer",
                "float",
                "regex",
                "json_schema"
            ],
            "x-enum-varnames": [
                "ValueValidatorParameterTypeNone",
                "ValueValidatorParameterTypeInteger",
                "ValueValidatorParameterTypeFloat",
                "ValueValidatorParameterTypeRegex",
                "ValueValidatorParameterTypeJsonSchema"
            ]
        },
...
        "valuetype.AllowedValidatorDto": {
            "type": "object",
            "required": [
                "parameterType",
                "validatorType"
            ],
            "properties": {
                "parameterType": {
                    "$ref": "#/definitions/validation.ValueValidatorParameterType"
                },
                "validatorType": {
                    "$ref": "#/definitions/db.ValueValidatorType"
                }
            }
        },
        "valuetype.ValidatorWithParameterTypeDto": {
            "type": "object",
            "required": [
                "errorText",
                "isBuiltIn",
                "parameter",
                "parameterType",
                "validatorType"
            ],
            "properties": {
                "errorText": {
                    "type": "string"
                },
                "isBuiltIn": {
                    "type": "boolean"
                },
                "parameter": {
                    "type": "string"
                },
                "parameterType": {
                    "$ref": "#/definitions/validation.ValueValidatorParameterType"
                },
                "validatorType": {
                    "$ref": "#/definitions/db.ValueValidatorType"
                }
            }
        },
        "valuetype.ValueTypeDto": {
            "type": "object",
            "required": [
                "allowedValidators",
                "id",
                "kind",
                "name",
                "validators"
            ],
            "properties": {
                "allowedValidators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valuetype.AllowedValidatorDto"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "$ref": "#/definitions/db.ValueTypeKind"
                },
                "name": {
                    "type": "string"
                },
                "validators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valuetype.ValidatorWithParameterTypeDto"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}